+++ C++/stack_queue.cpp
